+++ packages/apollo-server-plugin-response-cache/src/ApolloServerPluginResponseCache.ts
@@ -5,9 +5,10 @@ import type {
 import type {
   GraphQLRequestContext,
   GraphQLResponse,
+  CacheHint,
+  ValueOrPromise,
 } from 'apollo-server-types';
 import { KeyValueCache, PrefixingKeyValueCache } from 'apollo-server-caching';
-import type { CacheHint, ValueOrPromise } from 'apollo-server-types';
 import { CacheScope } from 'apollo-server-types';
 
 // XXX This should use createSHA from apollo-server-core in order to work on